+++ config/action-overflow.php
@@ -1,15 +1,14 @@
 <?php
 
-use Filament\Support\Enums\IconPosition;
-use Filament\Support\Icons\Heroicon;
+declare(strict_types=1);
 
 return [
     'primary_count' => 1,
     'label' => 'More',
-    'icon' => Heroicon::EllipsisVertical,
+    'icon' => 'heroicon-m-ellipsis-vertical',
     'color' => 'gray',
     'hidden_label' => false,
     'button' => true,
-    'icon_position' => IconPosition::After,
+    'icon_position' => 'after',
     'filter_unauthorized' => false,
 ];