+++ src/Macros/with_overflow_macro.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 use Filament\Actions\ActionGroup;
 use Harvirsidhu\FilamentActionOverflow\ActionOverflow;